% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrf.R
\name{ctrf}
\alias{ctrf}
\title{Clumping and Thresholding Random Forest (ctRF)}
\usage{
ctrf(
  G,
  CHR,
  POS,
  train_indices,
  test_indices,
  p_values,
  covariates,
  grid_type = "medium",
  n_cores = 4
)
}
\arguments{
\item{G}{Genotype matrix.}

\item{CHR}{Chromosome data.}

\item{POS}{Position data.}

\item{train_indices}{Training indices.}

\item{test_indices}{Test indices.}

\item{p_values}{P-values for SNPs.}

\item{covariates}{Covariates like age and sex.}

\item{grid_type}{Type of grid for hyperparameter tuning: 'light', 'medium', or 'intensive'.}

\item{n_cores}{Number of cores for parallel processing.}
}
\value{
A list containing the best model and predictions.
}
\description{
Clumping and Thresholding Random Forest (ctRF)
}
